Refactor controllers into playback and playlist
